% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge_kobo_data}
\alias{merge_kobo_data}
\title{Merging data files from deviating kobo tools}
\usage{
merge_kobo_data(folder, filename_pattern = ".", output_file = NULL,
  write_log = T, use_regex = F, add_file_reference = T,
  verbose = T)
}
\arguments{
\item{folder}{the path to the folder containing the files to be merged. The files must be in csv format.}

\item{filename_pattern}{a search term to find files that should be merged (inside the folder). You can use R style 'regex' patterns if you set use_regex=T}

\item{output_file}{full path to the file the merged dataset should be written to. If NULL (default), no output file is written.}

\item{write_log}{if TRUE, writes an html file with a log of the merge and subsequent checks.}

\item{use_regex}{if TRUE, searches the 'folder' directory for 'filename_pattern' using regex patterns. Defaults to FALSE.}

\item{add_file_reference}{If TRUE (default), adds a new column called `MERGED_FROM_FILE` containing the file name of each records source file.}

\item{verbose}{if TRUE writes messages to console during merge; set to FALSE to make quiet.}
}
\value{
the merged data frame.
}
\description{
Merging data files from deviating kobo tools
}
